test: move 7 Shifts service test files to Services/Shifts/
Per docs/plans/2026-05-12-section-align-shifts.md Axis 3.1. Canonical
home for Shifts service tests is tests/Humans.Application.Tests/Services/Shifts/.
