+++ tests/Humans.Application.Tests/Services/Shifts/ShiftDashboardMetricsTests.cs
@@ -21,7 +21,7 @@ using NodaTime.Testing;
 using NSubstitute;
 using Xunit;
 
-namespace Humans.Application.Tests.Services;
+namespace Humans.Application.Tests.Services.Shifts;
 
 public class ShiftDashboardMetricsTests : IDisposable
 {
