+++ tests/Humans.Application.Tests/Services/Shifts/ShiftUrgencyTests.cs
@@ -15,7 +15,7 @@ using NodaTime.Testing;
 using NSubstitute;
 using Xunit;
 
-namespace Humans.Application.Tests.Services;
+namespace Humans.Application.Tests.Services.Shifts;
 
 public class ShiftUrgencyTests
 {
